change how command line arguments are parsed
the 'argparse' module is now used to parse the command line arguments
which changes the arguments that are now used for the script.

this also fixes the previous issue where any error would result in
'invalid arguments' being printed out since all the contents of the
'main' function were in a 'try-except' block.